[runtime] Enhancement: Move pack_result to runtime
